ccfri 4764 moved ClosureDetailsDialog to components/closure
